Clean up docs, add release notes
mv src/file_parsing.jl into src/ClimaParams.jl

Rm Oceananigans example
