--- examples/OceananigansSingleColumnSimulation/Project.toml
@@ -1,9 +1,0 @@
-[deps]
-ClimaParams = "5c42b081-d73a-476f-9059-fd94b934656c"
-CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
-Oceananigans = "9e8cae18-63c1-5223-a75c-80ca9d6e9a09"
-SeawaterPolynomials = "d496a93d-167e-4197-9f49-d3af4ff8fe40"
-TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
-
-[compat]
-Oceananigans = "0.91"
