--- examples/OceananigansSingleColumnSimulation/default_parameters.toml
@@ -1,4 +1,0 @@
-[gravitational_acceleration]
-value = 9.81
-type = "float"
-
