--- examples/OceananigansSingleColumnSimulation/my_parameters.toml
@@ -1,4 +1,0 @@
-[gravitational_acceleration]
-value = 9.82
-type = "float"
-
