+++ test/runtests.jl
@@ -3,4 +3,4 @@ include("param_boxes.jl")
 include("types_from_file.jl")
 include("test_merge.jl")
 include("test_tags.jl")
-include("test_map.jl")
+# include("test_map.jl")
